requirements.txt
llama-cpp-python
transformers
torch
scikit-learn
nltk
PyQt6
python-docx
pdfminer.six
chardet